# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Sistema de Gestión de Clientes e Importaciones
  version: 1.1.0
  description: API Gateway para autenticación, clientes, solicitudes de importación y cotizaciones.
servers:
  - url: http://localhost:8080
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Token
paths:
  /health:
    get:
      summary: Verifica estado del API Gateway
      responses:
        '200':
          description: Servicio activo
  /auth/login:
    post:
      summary: Inicia sesión con usuario y contraseña
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username:
                  type: string
                  example: admin
                password:
                  type: string
                  example: admin123
      responses:
        '200':
          description: Sesión iniciada correctamente
        '401':
          description: Credenciales inválidas
  /auth/me:
    get:
      summary: Consulta el usuario autenticado
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Usuario autenticado
        '401':
          description: No autenticado
  /auth/logout:
    post:
      summary: Cierra la sesión actual
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Sesión cerrada
        '401':
          description: No autenticado
  /api/clientes:
    get:
      summary: Lista clientes
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Lista de clientes
        '401':
          description: No autenticado
    post:
      summary: Crea un cliente
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [nombre, email]
              properties:
                nombre:
                  type: string
                email:
                  type: string
                telefono:
                  type: string
                empresa:
                  type: string
      responses:
        '201':
          description: Cliente creado
        '401':
          description: No autenticado
  /api/clientes/{id}:
    get:
      summary: Obtiene un cliente por ID
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Cliente encontrado
        '401':
          description: No autenticado
    put:
      summary: Actualiza un cliente
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Cliente actualizado
        '401':
          description: No autenticado
  /api/solicitudes:
    get:
      summary: Lista solicitudes de importación
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Lista de solicitudes
        '401':
          description: No autenticado
    post:
      summary: Crea una solicitud de importación
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [cliente_id, producto, cantidad]
              properties:
                cliente_id:
                  type: integer
                producto:
                  type: string
                descripcion:
                  type: string
                pais_origen:
                  type: string
                cantidad:
                  type: integer
      responses:
        '201':
          description: Solicitud creada y evento publicado
        '401':
          description: No autenticado
  /api/solicitudes/{id}/estado:
    patch:
      summary: Actualiza estado de una solicitud
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [estado]
              properties:
                estado:
                  type: string
                  example: En proceso
      responses:
        '200':
          description: Estado actualizado
        '401':
          description: No autenticado
  /api/cotizaciones/calcular:
    post:
      summary: Calcula cotización mediante función serverless
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [valorProducto, envio, impuestosPorcentaje, margenPorcentaje]
              properties:
                valorProducto:
                  type: number
                envio:
                  type: number
                impuestosPorcentaje:
                  type: number
                margenPorcentaje:
                  type: number
      responses:
        '200':
          description: Cotización calculada
        '401':
          description: No autenticado